refactor(pages/learn): move ShadCN buttons and links to prod area
