+++ src/components/FeedbackWidget/FixedDot.tsx
@@ -2,11 +2,11 @@ import { forwardRef } from "react"
 import { useTranslation } from "next-i18next"
 import type { ButtonHTMLAttributes } from "react"
 
+import { Button } from "@/components/ui/buttons/Button"
+
 import { cn } from "@/lib/utils/cn"
 
 import { FeedbackGlyphIcon } from "../icons"
-
-import { Button } from "@/../tailwind/ui/buttons/Button"
 
 type FixedDotProps = ButtonHTMLAttributes<HTMLButtonElement> & {
   isExpanded: boolean
