+++ src/components/Nav/Mobile/HamburgerButton.tsx
@@ -6,10 +6,7 @@ import { cn } from "@/lib/utils/cn"
 
 import { HAMBURGER_BUTTON_ID } from "@/lib/constants"
 
-import {
-  Button,
-  type ButtonProps,
-} from "../../../../tailwind/ui/buttons/Button"
+import { Button, type ButtonProps } from "../../ui/buttons/Button"
 
 const hamburgerSvg =
   "M 2 13 l 10 0 l 0 0 l 10 0 M 4 19 l 8 0 M 12 19 l 8 0 M 2 25 l 10 0 l 0 0 l 10 0"
