+++ src/components/Nav/Mobile/LvlAccordion.tsx
@@ -6,8 +6,8 @@ import { cn } from "@/lib/utils/cn"
 import { trackCustomEvent } from "@/lib/utils/matomo"
 import { cleanPath } from "@/lib/utils/url"
 
-import { Button } from "../../../../tailwind/ui/buttons/Button"
-import { BaseLink } from "../../../../tailwind/ui/Link"
+import { Button } from "../../ui/buttons/Button"
+import { BaseLink } from "../../ui/Link"
 import type { Level, NavItem, NavSectionKey } from "../types"
 
 import ExpandIcon from "./ExpandIcon"
