+++ src/components/ui/buttons/ButtonTwoLines.tsx
@@ -2,7 +2,7 @@ import type { IconType } from "react-icons/lib"
 
 import { cn } from "@/lib/utils/cn"
 
-import { Stack } from "../../../src/components/ui/flex"
+import { Stack } from "../flex"
 
 import {
   Button,
